+++ src/common/Brush/Brush.story.tsx
@@ -16,13 +16,13 @@ import {
   LinearYAxisTickSeries,
   LinearYAxisTickLabel
 } from '../Axis';
-import { TooltipArea } from '../TooltipArea';
+import { TooltipArea } from '../Tooltip';
 import { ChartBrush } from './ChartBrush';
 import { ScatterPlot, ScatterSeries, ScatterPoint } from '../../ScatterPlot';
 import { timeDay } from 'd3-time';
 import { BarChart, BarSeries, Bar } from '../../BarChart';
 import { range } from 'd3-array';
-import { Tooltip } from '../../common/TooltipArea';
+import { Tooltip } from '../../common/Tooltip';
 import { GridlineSeries, Gridline, GridStripe } from '../Gridline';
 import { ChartZoomPan } from '../ZoomPan';