+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Master
 - [BREAKING] Update `meta` property on `ChartData` type to `metadata`
+- [BREAKING] Update `TooltipArea` namespace to just `Tooltip`
 - [Feature] Year Calendar Heatmaps
 - [Feature] Month Calendar Heatmaps
 - [Feature] Add ability to pass empty heatmap colors